fix(conf) set nginx_user to 'kong kong' if it exists
If we install and run Kong with everything default, whenever Kong tries
to write something to the Kong prefix, things will break and we will get
a permission denied error.

This applies to:

- DP nodes trying to write the config.json.gz file to the Kong prefix

- go pluginserver trying to write the go.lua file write to the Kong prefix

- Lua plugins trying to write files to the Kong prefix

And other cases. This is because, by default, the Kong prefix is owned
by root. When we do a apt-get install or yum install, the
/usr/local/kong/ prefix is owned by root by default. After that, if we
do a kong start without changing our $USER, the nginx master process
will be run as root and the nginx workers processes will be run as
nobody. It means that if we do everything default, the nginx workers
processes can’t write anything to the Kong prefix and we’ll get the
permission denied errors.

There are a few immediate workarounds for this:

- add the user nginx_user to the root group; or

- change the file system permissions to allow the nginx_user the
ability to write the cache file to /usr/local/kong with something like
`chown $nginx_user:$nginx_user /usr/local/kong/cache_config.gz; chmod 750 /usr/local/kong/config.cache.json.gz)`; or

- allow every user to write to /usr/local/kong with something like
chmod -R 755 /usr/local/kong (this is what was done in the ticket)

All of these workarounds are dangerous and hacky and we don’t document
any of these as recommended fixes. We should fix the root cause of the
issue. Here are the possibilities:

1) write files somewhere where any user has permissions, like /tmp. Maybe
even create a kong_prefix/tmp folder where it's explicitly said that
anyone has permissions over it (suggested by lluis).

2) set nginx_user to kong by default (instead of nobody). This will
require the non-root capability that we’re currently working on at
kong-build-tools: https://github.com/Kong/kong-build-tools/pull/339
(it will create a kong user and then chown and chmod the files and
directories owned by the kong package permissions to the kong user)

Out of those two I went with 2). I think 1) is a hack, we’d need to
repeat it everywhere we write files in the Kong prefix (Kong codebase,
Lua plugins, Go pluginserver, etc). It is also dangerous. IMO 2) is a
more general solution. Since nginx worker processes will be owned by
the kong user, things that try to write to the Kong prefix won't fail (the
`/usr/local/kong` prefix will be owned by the kong user by default upon
package installation after https://github.com/Kong/kong-build-tools/pull/339 is merged).

The idea of this PR is to fix the permission denied errors when running
Kong with the default settings without causing any breaking changes (if
the `kong` user does not exist, the user will be dropped to `nobody` as
it was before). If `nginx_user` is set to something else other than
`nobody` (default), we won't change anything.

I'm not sure where would be the ideal place to implement 2), my first
attempt is here in this commit with changes to the file
`kong/conf_loader/init.lua`.